Extract waveform audio file model
